fixed! wrote this out as I was debugging: I was declaring the module twice in the app, once in the routes angular   .module('Semio', [     'ngCookies',     'ngResource',     'ngRoute',     'ngSanitize',     'ngTouch'   ]) .config - route provider stuff
and once in app.js

   angular
        .module('Semio', [])
        .run(function(WordFactory) {
            WordFactory.getWords();
        });

It seems to make more sense to take it out of app.js because I ought to be able to call on it with angular.module(Semio) without having to instantiate it there, and I’m loading the module’s dependencies in the routes file. Angular Docs points out that this double instantiation can cause

Error: [$injector:unpr] Unknown provider: ServerUrlProvider <- ServerUrl <- WordFactory.

But when I take the module declaration out of app.js and simply call on it, I get

Uncaught Error: [$injector:nomod] Module 'Semio' is not available! You either misspelled the module name or forgot to load it. If registering a module ensure that you specify the dependencies as the second argument.

But the warning “Multiple modules with name "Semio" are being created and they will overwrite each other.” does go away. So, I’m definitely instantiating it twice, so the question is where it and its dependencies belong.

so next I moved the dependencies that were being loaded into the semio module into the instantiation in app.js, and accordingly switched the routes file to simply call on the module. This worked! I am no longer overwriting the module nor failing to load the dependencies - so in the end, both errors were active and related but different. It also seems to make more sense to inject the dependencies in the main app file, as that’s where I’m declaring constants and the run function.